DiscordBot/Hikari/requirements.txt
python-decouple
hikari
hikari-lightbulb
sentry-sdk
deepl
git+https://github.com/tandemdude/lightbulb-ext-filament.git